move MessageSink to own file
